+++ spark/src/main/scala/org/apache/comet/testing/FuzzDataGenerator.scala
@@ -44,6 +44,47 @@ object FuzzDataGenerator {
   val defaultBaseDate: Long =
     new SimpleDateFormat("YYYY-MM-DD hh:mm:ss").parse("3333-05-25 12:34:56").getTime
 
+  def generateSchema(options: SchemaGenOptions): StructType = {
+    val primitiveTypes = options.primitiveTypes
+    val dataTypes = ListBuffer[DataType]()
+    dataTypes.appendAll(primitiveTypes)
+
+    val arraysOfPrimitives = primitiveTypes.map(DataTypes.createArrayType)
+
+    if (options.generateStruct) {
+      dataTypes += StructType(
+        primitiveTypes.zipWithIndex.map(x => StructField(s"c${x._2}", x._1, nullable = true)))
+
+      if (options.generateArray) {
+        dataTypes += StructType(arraysOfPrimitives.zipWithIndex.map(x =>
+          StructField(s"c${x._2}", x._1, nullable = true)))
+      }
+    }
+
+    if (options.generateMap) {
+      dataTypes += MapType(DataTypes.IntegerType, DataTypes.StringType)
+    }
+
+    if (options.generateArray) {
+      dataTypes.appendAll(arraysOfPrimitives)
+
+      if (options.generateStruct) {
+        dataTypes += DataTypes.createArrayType(StructType(primitiveTypes.zipWithIndex.map(x =>
+          StructField(s"c${x._2}", x._1, nullable = true))))
+      }
+
+      if (options.generateMap) {
+        dataTypes += DataTypes.createArrayType(
+          MapType(DataTypes.IntegerType, DataTypes.StringType))
+      }
+    }
+
+    // generate schema using random data types
+    val fields = dataTypes.zipWithIndex
+      .map(i => StructField(s"c${i._2}", i._1, nullable = true))
+    StructType(fields.toSeq)
+  }
+
   def generateDataFrame(
       r: Random,
       spark: SparkSession,
@@ -178,6 +219,30 @@ object FuzzDataGenerator {
     }
   }
 }
+
+object SchemaGenOptions {
+  val defaultPrimitiveTypes: Seq[DataType] = Seq(
+    DataTypes.BooleanType,
+    DataTypes.ByteType,
+    DataTypes.ShortType,
+    DataTypes.IntegerType,
+    DataTypes.LongType,
+    DataTypes.FloatType,
+    DataTypes.DoubleType,
+    DataTypes.createDecimalType(10, 2),
+    DataTypes.createDecimalType(36, 18),
+    DataTypes.DateType,
+    DataTypes.TimestampType,
+    DataTypes.TimestampNTZType,
+    DataTypes.StringType,
+    DataTypes.BinaryType)
+}
+
+case class SchemaGenOptions(
+    generateArray: Boolean = false,
+    generateStruct: Boolean = false,
+    generateMap: Boolean = false,
+    primitiveTypes: Seq[DataType] = SchemaGenOptions.defaultPrimitiveTypes)
 
 case class DataGenOptions(
     allowNull: Boolean = true,
