+++ spark/src/main/scala/org/apache/comet/testing/ParquetGenerator.scala
@@ -19,11 +19,10 @@
 
 package org.apache.comet.testing
 
-import scala.collection.mutable.ListBuffer
 import scala.util.Random
 
 import org.apache.spark.sql.{SaveMode, SparkSession}
-import org.apache.spark.sql.types.{DataType, DataTypes, MapType, StructField, StructType}
+import org.apache.spark.sql.types.{DataType, StructType}
 
 object ParquetGenerator {
 
@@ -34,7 +33,13 @@ object ParquetGenerator {
       filename: String,
       numRows: Int,
       options: ParquetGeneratorOptions): Unit = {
-    val schema = generateSchema(options)
+
+    val schemaGenOptions = SchemaGenOptions(
+      generateArray = options.generateArray,
+      generateStruct = options.generateStruct,
+      generateMap = options.generateMap,
+      primitiveTypes = options.primitiveTypes)
+    val schema = FuzzDataGenerator.generateSchema(schemaGenOptions)
 
     val dataGenOptions = DataGenOptions(
       allowNull = options.allowNull,
@@ -56,46 +61,6 @@ object ParquetGenerator {
     df.write.mode(SaveMode.Overwrite).parquet(filename)
   }
 
-  private def generateSchema(options: ParquetGeneratorOptions): StructType = {
-    val primitiveTypes = options.primitiveTypes
-    val dataTypes = ListBuffer[DataType]()
-    dataTypes.appendAll(primitiveTypes)
-
-    val arraysOfPrimitives = primitiveTypes.map(DataTypes.createArrayType)
-
-    if (options.generateStruct) {
-      dataTypes += StructType(
-        primitiveTypes.zipWithIndex.map(x => StructField(s"c${x._2}", x._1, nullable = true)))
-
-      if (options.generateArray) {
-        dataTypes += StructType(arraysOfPrimitives.zipWithIndex.map(x =>
-          StructField(s"c${x._2}", x._1, nullable = true)))
-      }
-    }
-
-    if (options.generateMap) {
-      dataTypes += MapType(DataTypes.IntegerType, DataTypes.StringType)
-    }
-
-    if (options.generateArray) {
-      dataTypes.appendAll(arraysOfPrimitives)
-
-      if (options.generateStruct) {
-        dataTypes += DataTypes.createArrayType(StructType(primitiveTypes.zipWithIndex.map(x =>
-          StructField(s"c${x._2}", x._1, nullable = true))))
-      }
-
-      if (options.generateMap) {
-        dataTypes += DataTypes.createArrayType(
-          MapType(DataTypes.IntegerType, DataTypes.StringType))
-      }
-    }
-
-    // generate schema using random data types
-    val fields = dataTypes.zipWithIndex
-      .map(i => StructField(s"c${i._2}", i._1, nullable = true))
-    StructType(fields.toSeq)
-  }
 }
 
 /** Schema and Data generation options */
@@ -106,18 +71,4 @@ case class ParquetGeneratorOptions(
     generateArray: Boolean = false,
     generateStruct: Boolean = false,
     generateMap: Boolean = false,
-    primitiveTypes: Seq[DataType] = Seq(
-      DataTypes.BooleanType,
-      DataTypes.ByteType,
-      DataTypes.ShortType,
-      DataTypes.IntegerType,
-      DataTypes.LongType,
-      DataTypes.FloatType,
-      DataTypes.DoubleType,
-      DataTypes.createDecimalType(10, 2),
-      DataTypes.createDecimalType(36, 18),
-      DataTypes.DateType,
-      DataTypes.TimestampType,
-      DataTypes.TimestampNTZType,
-      DataTypes.StringType,
-      DataTypes.BinaryType))
+    primitiveTypes: Seq[DataType] = SchemaGenOptions.defaultPrimitiveTypes)
